Improve custom-properties-allowed-list plugin
