+++ .eslintignore
@@ -4,7 +4,3 @@ node_modules
 /build
 /build-internal
 /src/styles/polaris-tokens
-
-index.esnext
-index.js
-index.mjs
