+++ loom.config.ts
@@ -26,8 +26,10 @@ import type {} from '@shopify/loom-plugin-jest';
 // eslint-disable-next-line import/no-default-export
 export default createPackage((pkg) => {
   pkg.entry({root: './src/index.ts'});
+  pkg.entry({name: 'token-list', root: './src/token-list.ts'});
   pkg.use(
     buildLibrary({
+      rootEntrypoints: false,
       jestTestEnvironment: 'jsdom',
       targets: 'extends @shopify/browserslist-config, node 12.20',
       commonjs: true,
