+++ scripts/stylelint/config/index.js
@@ -1,9 +1,23 @@
 const path = require('path');
 
-require('ts-node').register({
-  project: path.join(__dirname, '../../tsconfig.json'),
-});
+const {tokenList} = require('../../../build/cjs/token-list');
 
-const {stylelintConfig} = require('./config')
+/**
+ * Allowed custom property names in Polaris component styles.
+ */
+const polarisComponentCustomProperties = /--pc-.+/;
 
-module.exports = stylelintConfig;
+module.exports = {
+  plugins: [path.join(__dirname, '../plugins/custom-properties-allowed-list')],
+  rules: {
+    '@shopify/custom-properties-allowed-list': {
+      allowedProperties: [
+        '--polaris-version-number',
+        polarisComponentCustomProperties,
+      ],
+      allowedValues: {
+        '/.+/': [polarisComponentCustomProperties, ...tokenList],
+      },
+    },
+  },
+};
