--- scripts/stylelint/plugins/custom-properties-allowed-list/utils/index.ts
@@ -1,4 +1,0 @@
-export * as vendor from './vendor';
-export * from './matchesStringOrRegExp';
-export * from './isCustomProperty';
-export * from './validateTypes';
