--- scripts/stylelint/plugins/custom-properties-allowed-list/utils/isCustomProperty.ts
@@ -1,8 +1,0 @@
-/**
- * Check whether a property is a custom one
- * @param {string} property
- * @returns {boolean}
- */
-export function isCustomProperty(property: string) {
-  return property.startsWith('--');
-}
