--- scripts/stylelint/plugins/custom-properties-allowed-list/utils/validateTypes.ts
@@ -1,35 +1,0 @@
-/**
- * Checks if the value is a boolean or a Boolean object.
- * @param {unknown} value
- * @returns {value is boolean}
- */
-export function isBoolean(value: unknown) {
-  return typeof value === 'boolean' || value instanceof Boolean;
-}
-
-/**
- * Checks if the value is a number or a Number object.
- * @param {unknown} value
- * @returns {value is number}
- */
-export function isNumber(value: unknown) {
-  return typeof value === 'number' || value instanceof Number;
-}
-
-/**
- * Checks if the value is a RegExp object.
- * @param {unknown} value
- * @returns {value is RegExp}
- */
-export function isRegExp(value: unknown) {
-  return value instanceof RegExp;
-}
-
-/**
- * Checks if the value is a string or a String object.
- * @param {unknown} value
- * @returns {value is string}
- */
-export function isString(value: unknown) {
-  return typeof value === 'string' || value instanceof String;
-}
