--- scripts/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "extends": "@shopify/typescript-configs/base.json",
-
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2019",
-    "rootDir": "../",
-    "outDir": "./dist",
-    "noEmit": true,
-  },
-  "include": ["./**/*"],
-  "ts-node": {
-    // It is faster to skip typechecking.
-    // Remove if you want ts-node to do typechecking.
-    "transpileOnly": true,
-
-    "files": true,
-
-    "compilerOptions": {},
-  }
-}
